+++ tgbotapi.core/src/commonMain/kotlin/dev/inmo/tgbotapi/types/message/content/media/AudioMediaGroupContent.kt
@@ -3,7 +3,7 @@ package dev.inmo.tgbotapi.types.message.content.media
 import dev.inmo.tgbotapi.types.InputMedia.AudioMediaGroupMemberInputMedia
 import dev.inmo.tgbotapi.types.files.AudioFile
 
-interface AudioMediaGroupContent : TextedMediaGroupMediaContent {
+interface AudioMediaGroupContent : MediaGroupContent {
     override val media: AudioFile
 
     override fun toMediaGroupMemberInputMedia(): AudioMediaGroupMemberInputMedia